scraper/scraper.py
import requests
from bs4 import BeautifulSoup
from database.models import db, Property
from datetime import datetime

class RealEstateScraper:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }

    def scrape_properties(self, url):
        """
        Scrape property listings from the given URL
        This is a placeholder method - implement specific scraping logic based on target website
        """
        try:
            response = requests.get(url, headers=self.headers)
            response.raise_for_status()
            
            # Implementation will depend on the specific website structure
            # This is just a skeleton
            
            soup = BeautifulSoup(response.text, 'html.parser')
            # Add specific scraping logic here
            
        except Exception as e:
            print(f"Error scraping properties: {str(e)}")
            return []

    def save_to_db(self, property_data):
        """
        Save scraped property data to database
        """
        try:
            property_instance = Property(**property_data)
            db.session.add(property_instance)
            db.session.commit()
        except Exception as e:
            print(f"Error saving to database: {str(e)}")
            db.session.rollback()
