requirements.txt
Flask
SQLAlchemy
psycopg2-binary
beautifulsoup4
requests
python-dotenv
Flask-SQLAlchemy
